remove old LMS1xx library
