+++ CMakeLists.txt
@@ -12,17 +12,16 @@ add_compile_options(-std=c++11)
 find_package(console_bridge REQUIRED)
 include_directories(include SYSTEM ${console_bridge_INCLUDE_DIRS})
 
+# CoLaA Library abstracting protocol and implementing LMS1xx communication
 add_library(CoLaA src/colaa.cpp src/parse_helpers.cpp)
 target_link_libraries(CoLaA ${console_bridge_LIBRARIES})
 
-add_library(LMS1xx src/LMS1xx.cpp)
-target_link_libraries(LMS1xx ${console_bridge_LIBRARIES})
+# Specialisations for LMS5xx series scanners
+add_library(LMS5xx src/lms5xx.cpp)
+target_link_libraries(LMS5xx CoLaA ${console_bridge_LIBRARIES})
 
 add_library(MRS1000 src/MRS1000.cpp)
 target_link_libraries(MRS1000 CoLaA ${console_bridge_LIBRARIES})
-
-add_library(LMS5xx src/lms5xx.cpp)
-target_link_libraries(LMS5xx CoLaA ${console_bridge_LIBRARIES})
 
 
 # Regular catkin package follows.
@@ -30,8 +29,10 @@ find_package(catkin REQUIRED COMPONENTS roscpp sensor_msgs)
 catkin_package(CATKIN_DEPENDS roscpp)
 
 include_directories(include SYSTEM ${catkin_INCLUDE_DIRS})
+
 add_executable(LMS1xx_node src/LMS1xx_node.cpp)
-target_link_libraries(LMS1xx_node LMS1xx ${catkin_LIBRARIES})
+target_link_libraries(LMS1xx_node CoLaA ${catkin_LIBRARIES})
+add_dependencies(LMS1xx_node CoLaA)
 
 add_executable(MRS1000_node src/MRS1000_node.cpp)
 target_link_libraries(MRS1000_node MRS1000 ${catkin_LIBRARIES})
@@ -41,11 +42,7 @@ add_executable(LMS5xx_node src/lms5xx_node.cpp)
 target_link_libraries(LMS5xx_node LMS5xx ${catkin_LIBRARIES})
 add_dependencies(LMS5xx_node LMS5xx)
 
-add_executable(TestNode src/testnode.cpp)
-target_link_libraries(TestNode CoLaA ${catkin_LIBRARIES})
-add_dependencies(TestNode CoLaA)
-
-install(TARGETS LMS1xx LMS1xx_node MRS1000 MRS1000_node
+install(TARGETS CoLaA LMS5xx MRS1000 LMS1xx_node LMS5xx_node MRS1000_node
   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
